fix(types): resolve 9 mypy errors caught by CI + newer stubs
- geo.py: wrap curvature-radius returns in float() so mypy sees the
  declared return type (numpy operations returned Any).
- earth/layered_1d.py: narrow compute_impedance return type to
  ScalarImpedance (covariant return over the ABC's Impedance),
  matching the docstring and unblocking .Z_values access downstream.
- earth/structured_2d.py: no code change; the narrower Layered1D
  return type now type-checks.
- config.py: silence yaml import-untyped (types-PyYAML not in deps).
- uq/uncertain.py: type: ignore on numpy calls with unbounded T.
- network/pipeline.py: cast analytic-DSTL return to np.ndarray.

All 100 tests still pass; ruff + mypy now both clean locally.

+++ src/geopulse/geo.py
@@ -67,7 +67,9 @@ def meridian_radius_m(lat_deg: float) -> float:
     of the ellipsoid along a meridian (north-south).
     """
     sin_phi = np.sin(lat_deg * DEG_TO_RAD)
-    return WGS84_A_M * (1.0 - WGS84_E2) / (1.0 - WGS84_E2 * sin_phi * sin_phi) ** 1.5
+    return float(
+        WGS84_A_M * (1.0 - WGS84_E2) / (1.0 - WGS84_E2 * sin_phi * sin_phi) ** 1.5
+    )
 
 
 def prime_vertical_radius_m(lat_deg: float) -> float:
@@ -77,7 +79,7 @@ def prime_vertical_radius_m(lat_deg: float) -> float:
     perpendicular to a meridian, at latitude ``φ``.
     """
     sin_phi = np.sin(lat_deg * DEG_TO_RAD)
-    return WGS84_A_M / np.sqrt(1.0 - WGS84_E2 * sin_phi * sin_phi)
+    return float(WGS84_A_M / np.sqrt(1.0 - WGS84_E2 * sin_phi * sin_phi))
 
 
 def latlon_to_local_xy_spherical_m(
